cloud_backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# AI and ML (essential only)
google-generativeai==0.3.2
pillow==10.0.1
numpy==1.24.3

# Database
asyncpg==0.29.0
redis==5.0.1

# Utilities
python-dotenv==1.0.0
qrcode[pil]==7.4.2
aiofiles==23.2.1
requests==2.31.0

# Essential for image processing (lightweight)
opencv-python-headless==4.8.1.78
# Note: YOLO disabled for faster deployment - Gemini AI will handle classification